requirements.txt
torch
torchvision
gradio
Pillow
pandas
